requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
pytest>=7.0
pytest-django>=4.0
